fix: mapping result scheme
